[PRMT-4524] Moved configs to configuration package.
