+++ src/integration/java/uk/nhs/prm/repo/ehrtransferservice/configuration/ScheduledConfig.java
@@ -1,4 +1,4 @@
-package uk.nhs.prm.repo.ehrtransferservice;
+package uk.nhs.prm.repo.ehrtransferservice.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
